Crea el componente FAQCard con sus estilos y se muestran dentro de la página de preguntas y respuestas
