+++ src/styles/FAQCard.css
@@ -1,0 +1,40 @@
+.FAQCardContainer { 
+    display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    max-width: 80%;
+    overflow: hidden;
+    gap: 0;
+}
+
+.questionContainer { 
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    max-height: 4vh;
+    background-color: #000000;
+    padding: 8px 12px 8px 12px;
+}
+
+.question { 
+    font-size: 24px;
+    text-wrap: wrap;
+    background-color: #000000;
+    color: #FFFFFF;
+    font-weight: bold;
+}
+
+.circleButton { 
+    min-width: 32px;
+    min-width: 32px;
+}
+
+.answerContainer { 
+    background-color: #F1F1F1;
+    padding: 0px 12px;
+}
+
+.answer {
+    font-size: 20px;
+    font-weight: 500;
+}
